applying comments for pr
