+++ dataset/__init__.py
@@ -1,7 +1,5 @@
 from .dataset import Dataset
 from .dataset import AuxTables
 from .dataset import CellStatus
-from .table import Source
-from .table import Table
 
-__all__ = ['Dataset', 'AuxTables', 'CellStatus', 'Source', 'Table']
+__all__ = ['Dataset', 'AuxTables', 'CellStatus']
