+++ detect/__init__.py
@@ -2,7 +2,7 @@ from .detect import DetectEngine
 from .detector import Detector
 from .nulldetector import NullDetector
 from .violationdetector import ViolationDetector
-from .fixeddetector import FixedDetector
+from .errorloaderdetector import ErrorsLoaderDetector
 
 __all__ = ['DetectEngine', 'Detector', 'NullDetector', 
-            'ViolationDetector', 'FixedDetector']
+            'ViolationDetector', 'ErrorsLoaderDetector']
